backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
supabase==2.9.0

# Environment
python-dotenv==1.0.1

# Validation
pydantic==2.9.2
pydantic-settings==2.6.0

# CORS
python-jose[cryptography]==3.3.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# ML/Training (Phase 6 - Optional, install when needed)
# Uncomment these when you start Phase 6 model training:
# torch>=2.6.0  # Latest version compatible with Python 3.13
# tensorboard>=2.18.0
# onnx>=1.17.0
# tensorflowjs (install via npm when needed)
